+++ packages/common/types.ts
@@ -1,7 +1,17 @@
+export const reservedPageNames = ["page", "attachment", "plug"];
+
 export type PageMeta = {
   name: string;
   lastModified: number;
   lastOpened?: number;
+  perm: "ro" | "rw";
+};
+
+export type AttachmentMeta = {
+  name: string;
+  contentType: string;
+  lastModified: number;
+  size: number;
   perm: "ro" | "rw";
 };