+++ packages/web/editor_paste.ts
@@ -1,4 +1,5 @@
-import { ViewPlugin, ViewUpdate } from "@codemirror/view";
+import { EditorView, ViewPlugin, ViewUpdate } from "@codemirror/view";
+import { Space } from "@silverbulletmd/common/spaces/space";
 import { createImportSpecifier } from "typescript";
 
 const urlRegexp =
@@ -43,3 +44,51 @@ export const pasteLinkExtension = ViewPlugin.fromClass(
     }
   }
 );
+
+export function pasteAttachmentExtension(space: Space) {
+  return EditorView.domEventHandlers({
+    paste: (event: ClipboardEvent, editor) => {
+      let payload = [...event.clipboardData!.items];
+
+      if (!payload.length || payload.length === 0) {
+        return false;
+      }
+      let file = payload.find((item) => item.kind === "file");
+      if (!file) {
+        return false;
+      }
+      const fileType = file.type;
+      Promise.resolve()
+        .then(async () => {
+          let data = await file!.getAsFile()?.arrayBuffer();
+          let ext = fileType.split("/")[1];
+          let fileName = new Date()
+            .toISOString()
+            .split(".")[0]
+            .replace("T", "_")
+            .replaceAll(":", "-");
+          let finalFileName = prompt(
+            "File name for pasted attachment",
+            `${fileName}.${ext}`
+          );
+          if (!finalFileName) {
+            return;
+          }
+          await space.writeAttachment(finalFileName, data!);
+          let attachmentMarkdown = `[${finalFileName}](attachment/${finalFileName})`;
+          if (fileType.startsWith("image/")) {
+            attachmentMarkdown = `![](attachment/${finalFileName})`;
+          }
+          editor.dispatch({
+            changes: [
+              {
+                insert: attachmentMarkdown,
+                from: editor.state.selection.main.from,
+              },
+            ],
+          });
+        })
+        .catch(console.error);
+    },
+  });
+}