+++ packages/web/service_worker.ts
@@ -35,8 +35,10 @@ self.addEventListener("fetch", (event: any) => {
         return response;
       } else {
         if (
-          parsedUrl.pathname !== "/fs" &&
-          !parsedUrl.pathname.startsWith("/fs/") &&
+          parsedUrl.pathname !== "/page" &&
+          !parsedUrl.pathname.startsWith("/page/") &&
+          parsedUrl.pathname !== "/attachment" &&
+          !parsedUrl.pathname.startsWith("/attachment/") &&
           !parsedUrl.pathname.startsWith("/plug/")
         ) {
           return cache.match("/index.html");